move proxies, add audio
